Refactor colorpickers interfaces & types
