+++ packages/colorpickers/src/elements/ColorpickerDialog/index.tsx
@@ -12,74 +12,20 @@ import React, {
   Children,
   cloneElement,
   forwardRef,
-  ReactElement,
-  HTMLAttributes
+  ReactElement
 } from 'react';
 import PropTypes from 'prop-types';
-import { Modifier } from 'react-popper';
 import { Button } from '@zendeskgarden/react-buttons';
-import { GARDEN_PLACEMENT } from '@zendeskgarden/react-modals';
 import { composeEventHandlers } from '@zendeskgarden/container-utilities';
 import Chevron from '@zendeskgarden/svg-icons/src/16/chevron-down-stroke.svg';
-import { Colorpicker, IColorpickerProps } from '../Colorpicker';
-import { IColor } from '../../utils/types';
+import { Colorpicker } from '../Colorpicker';
 import {
   StyledButton,
   StyledButtonPreview,
   StyledTooltipModal,
   StyledTooltipBody
 } from '../../styled';
-
-export interface IColorpickerDialogChanges {
-  isOpen?: boolean;
-}
-
-export interface IColorpickerDialogProps extends IColorpickerProps {
-  /**
-   * Handles close actions. Can be triggered from the backdrop.
-   *
-   * @param {Object} color A color picker state
-   */
-  onClose?: (color: IColor) => void;
-  /** Adjusts the placement of the color dialog */
-  placement?: GARDEN_PLACEMENT;
-  /** Disables the color dialog button */
-  disabled?: boolean;
-  /**
-   * Modifies [Popper instance](https://popper.js.org/docs/v2/modifiers/) to customize positioning logic
-   */
-  popperModifiers?: Partial<Modifier<any, any>>[];
-  /**
-   * Sets the `z-index` of the color dialog
-   */
-  zIndex?: number;
-  /**
-   * Adds an arrow to the color dialog
-   */
-  hasArrow?: boolean;
-  /**
-   * Animates the color dialog
-   */
-  isAnimated?: boolean;
-  /**
-   * Opens the dialog in a controlled color picker dialog
-   */
-  isOpen?: boolean;
-  /**
-   * Applies inset `box-shadow` styling on focus
-   */
-  focusInset?: boolean;
-  /**
-   * Passes HTML attributes to the color dialog button element
-   */
-  buttonProps?: HTMLAttributes<HTMLButtonElement>;
-  /**
-   * Handles dialog changes
-   *
-   * @param {Object} changes The changed dialog state
-   */
-  onDialogChange?: (changes: IColorpickerDialogChanges) => void;
-}
+import { IColor, IColorpickerDialogProps } from '../../types';
 
 /**
  * @extends HTMLAttributes<HTMLDivElement>
